\name{turf_ex_data}
\docType{data}
\alias{turf_ex_data}
\title{Example data set for TURF}
\description{Data used in examples and to test functionality of turfR.
}
\usage{turf_ex_data}
\format{A data frame containing 180 observations (rows) and 12 variables (columns), all numeric. Columns are \code{respid}:
  a unique row identifier; \code{weight}: an individual-level weight variable that sums to 180 (i.e., the number of rows,
  mean = 1); and \code{item_1} to \code{item_10}: 10 binary-coded (0/1) variables representing TURF data for 10 items.
}
\source{Generated using \code{runif} and related functions.}
\keyword{datasets}