% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{turf_ex_output}
\alias{turf_ex_output}
\title{turf_ex_output: Example data set of an output of the turf function}
\format{
A data frame containing 45 observations (rows) and 13 variables (columns), all numeric:
\describe{
\item{combo}{combination}
\item{rchX}{Numeric value representing the reach}
\item{frqX}{Numeric value representing the frequency}
\item{1}{Binary-coded (0/1) variable representing data for item 1}
\item{2}{Binary-coded (0/1) variable representing data for item 2}
\item{3}{Binary-coded (0/1) variable representing data for item 3}
\item{4}{Binary-coded (0/1) variable representing data for item 4}
\item{5}{Binary-coded (0/1) variable representing data for item 5}
\item{6}{Binary-coded (0/1) variable representing data for item 6}
\item{7}{Binary-coded (0/1) variable representing data for item 7}
\item{8}{Binary-coded (0/1) variable representing data for item 8}
\item{9}{Binary-coded (0/1) variable representing data for item 9}
\item{10}{Binary-coded (0/1) variable representing data for item 10}
}
}
\usage{
turf_ex_output
}
\description{
An example of turf output data used to test functionality of turfR.
}
\keyword{datasets}
