\name{turfR-package}
\alias{turfR-package}
\docType{package}
\title{TURF Analysis for R}
\description{
    Package for analyzing TURF (Total Unduplicated Reach and Frequency) data in R. 
    No looping in TURF algorithm results in fast processing times. Allows for individual-level 
	weights, depth specification, and user-truncated combination set(s). Allows user to substitute 
	Monte Carlo simulated combination set(s) after set(s) exceed a user-specified limit.
	
	See \code{help(turf)} for additional details and references.
}