% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turfR_0.8-7.R
\name{turf.combos}
\alias{turf.combos}
\title{Generate combination sets to be used in TURF analysis, formatted for use in turf. Can be used to generate all possible combinations of n choose k or Monte Carlo simulated subsets of combinations involving the same n items.}
\usage{
turf.combos(n, k, ...)
}
\arguments{
\item{n}{- Required. Scalar indicating the number of items to be included in the TURF algorithm; 0 < n < (ncol(data) - 1). Non-integer values are coerced using floor(n).}

\item{k}{- Required. Vector of length 1 to n containing any values 1 to n indicating the combination sizes to be evaluated by the TURF algorithm. Non-integer values are coerced using floor(k).}

\item{...}{- Optional. Additional arguments controlling behavior and output of the TURF algorithm. See turf.args. Arguments indicated here must match named arguments in turf.args.}
}
\value{
combination set formatted for use in turf
}
\description{
Generate combination sets to be used in TURF analysis, formatted for use in turf. Can be used to generate all possible combinations of n choose k or Monte Carlo simulated subsets of combinations involving the same n items.
}
