% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turfR_0.8-7.R
\name{turf}
\alias{turf}
\title{Function for undertaking turf analysis}
\usage{
turf(data, n, k, combos, ...)
}
\arguments{
\item{data}{\itemize{
\item Required. Literal character string representing name of a file in the working directory readable using read.table(data, header=TRUE), or name of a data frame or matrix in R containing TURF data. Rows are individuals (respondents). Columns are (1) respondent identifier, (2) a weight variable, and a minimum of n columns containing only zeroes and ones, each representing an individual item in the TURF algorithm. Respondent identifiers need not be unique and weights need not sum to the total number of rows. In the absence of any weight variable, substitute a column of ones. Ones in the remaining columns indicate that the reach criterion was met for a given item by a given individual. Values other than zero or one in these columns (including NA) trigger an error. data may contain more than n + 2 columns, but any columns in addition to that number will be ignored.
}}

\item{n}{\itemize{
\item Required. Scalar indicating the number of items to be included in the TURF algorithm; 0 < n < (ncol(data) - 1). Non-integer values are coerced using floor(n).
}}

\item{k}{\itemize{
\item Required. Vector of length 1 to n containing any values 1 to n indicating the combination sizes to be evaluated by the TURF algorithm. Non-integer values are coerced using floor(k).
}}

\item{combos}{\itemize{
\item Optional. List of combination sets to be evaluated by the TURF algorithm, such as that generated by turf.combos. Individual combination sets are p x n matrices, containing only zeroes and ones indicating items to be included in a given combination. Rows (p) correspond to combinations evaluated; columns (n) correspond to items, and must be in the same order as the items columns in data. Each ith element of combos should contain combinations of the size specified in the ith position of k; length(combos) must be equal to length(k). See "details" for additional information on usage.
}}

\item{...}{\itemize{
\item Optional. Additional arguments controlling behavior and output of the TURF algorithm. See turf.args. Arguments indicated here must match named arguments in turf.args.
}}
}
\value{
turf dataframe
}
\description{
Function for undertaking turf analysis
}
